+++ src/client/cmd.rs
@@ -1,3 +1,5 @@
+use crate::shell::interactive_shell_cmd_line;
+
 /// Prepare a command list for executing the given shell line.
 ///
 /// This will always call out to bash.
@@ -6,4 +8,14 @@ pub fn shell_to_cmd(shell_cmd: &str) -> Vec<String> {
         .into_iter()
         .map(String::from)
         .collect()
+}
+
+/// Prepare a command list for executing the given shell line in an interactive
+/// login shell.
+pub fn shell_to_cmd_interactive(shell_cmd: &str) -> Vec<String> {
+    let mut interactive_shell_cmd =
+        interactive_shell_cmd_line(false).expect("Unexpected shell");
+    interactive_shell_cmd.push(String::from("-c"));
+    interactive_shell_cmd.push(String::from(shell_cmd));
+    interactive_shell_cmd
 }