+++ src/dependency.rs
@@ -605,6 +605,7 @@ mod test {
                 None,
                 None,
                 None,
+                false,
             )),
         }
     }